asim: fix shorter than expected replica change delay
Previously, we used integer division to compute the delay for adding a replica,
which led to shorter than expected delays. We previously attempted to fix this
by switching to float division but didn't notice we are later casting to
time.Duration which truncated the result. This commit fixes the issue by casting
time.Second to a float instead.

Before:
```
I250709 03:24:05.016545 47 kv/kvserver/asim/op/settings.go:158 ⋮ [-] 54117  estimatedTimeToAddReplica: 0.80, s.RebalancingSnapshotRate: 33554432, base: 100000000
I250709 03:24:05.016549 47 kv/kvserver/asim/op/settings.go:161 ⋮ [-] 54118  calculated replica change delay: ‹100ms› for range size 26843545
```

After:
```
I250709 03:27:53.914048 43 kv/kvserver/asim/op/settings.go:159 ⋮ [-] 11096  estimatedTimeToAddReplica: 0.80, s.RebalancingSnapshotRate: 33554432, base: 100000000
I250709 03:27:53.914053 43 kv/kvserver/asim/op/settings.go:162 ⋮ [-] 11097  calculated replica change delay: ‹899.999982ms› for range size 26843545
```

See slack https://cockroachlabs.slack.com/archives/C048HDZJSAY/p1751935031605719 for more context.

Epic: none
Release note: none

+++ pkg/kv/kvserver/asim/config/settings.go
@@ -163,7 +163,9 @@ func (s *SimulationSettings) ReplicaChangeDelayFn() func(rangeSize int64, add bo
 		delay := s.ReplicaChangeBaseDelay
 		if add {
 			estimatedTimeToAddReplica := float64(rangeSize) / float64(s.RebalancingSnapshotRate)
-			delay += time.Duration(estimatedTimeToAddReplica) * time.Second
+			// To avoid truncating the decimal part of the result, cast to Duration
+			// after float multiplication.
+			delay += time.Duration(estimatedTimeToAddReplica * float64(time.Second))
 		}
 		return delay
 	}
